# ui/public/doc/api/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Бот-помощник по организации работы команды и увеличения ее эффективности.\nПозволяет выполнять определенные действия по расписанию.",
    "title": "PolkovnikBot API.",
    "contact": {
      "name": "Ladygin Sergey",
      "email": "sladygin@updev.ru"
    },
    "version": "1.1"
  },
  "basePath": "/api",
  "paths": {
...
    "/team/{teamId}/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Список участников команды.",
        "operationId": "UserList",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TeamId",
            "description": "ID команды",
            "name": "teamId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/UserList"
          }
        }
      },
      "post": {
        "description": "Создание нового участника",
        "tags": [
          "Users"
        ],
        "operationId": "UserAdd",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TeamId",
            "description": "ID команды",
            "name": "teamId",
            "in": "path",
            "required": true
          },
          {
            "x-go-name": "Body",
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/User"
          },
          "400": {
            "$ref": "#/responses/ResponseError"
          }
        }
      }
    },
    "/team/{teamId}/users/{userId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Информация об участнике команды.",
        "operationId": "UserGet",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TeamId",
            "description": "ID команды",
            "name": "teamId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "UserId",
            "description": "ID Пользователя",
            "name": "userId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/User"
          },
          "404": {
            "$ref": "#/responses/ResponseError"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Удаление участника команды.",
        "operationId": "UserDelete",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TeamId",
            "description": "ID команды",
            "name": "teamId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "UserId",
            "description": "ID Пользователя",
            "name": "userId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseSuccess"
          },
          "404": {
            "$ref": "#/responses/ResponseError"
          }
        }
      },
      "patch": {
        "description": "Изменение участника",
        "tags": [
          "Users"
        ],
        "operationId": "UserEdit",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TeamId",
            "description": "ID команды",
            "name": "teamId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "UserId",
            "description": "ID Пользователя",
            "name": "userId",
            "in": "path",
            "required": true
          },
          {
            "x-go-name": "Body",
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/User"
          },
          "400": {
            "$ref": "#/responses/ResponseError"
          },
          "404": {
            "$ref": "#/responses/ResponseError"
          }
        }
      }
    }
  },
  "definitions": {
...
    "User": {
      "description": "Участник команды",
      "type": "object",
      "properties": {
        "active": {
          "type": "boolean",
          "x-go-name": "Active"
        },
        "id": {
          "type": "string",
          "x-go-name": "Id"
        },
        "login": {
          "type": "string",
          "x-go-name": "Login"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "nickname": {
          "type": "string",
          "x-go-name": "NickName"
        },
        "weekend": {
          "$ref": "#/definitions/Weekend"
        }
      },
      "x-go-name": "userResponseItem",
      "x-go-package": "polkovnik/api"
    },
    "Users": {
      "description": "Список участников команды",
      "type": "object",
      "properties": {
        "result": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User"
          },
          "x-go-name": "Result"
        }
      },
      "x-go-name": "userResponseList",
      "x-go-package": "polkovnik/api"
    },
    "Weekend": {
      "description": "Выходные",
      "type": "object",
      "properties": {
        "intervals": {
          "description": "Интервалы дат",
          "type": "array",
          "items": {
            "$ref": "#/definitions/WeekendInterval"
          },
          "x-go-name": "Intervals"
        },
        "week_days": {
          "description": "Выходные дни",
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "WeekDays"
        }
      },
      "x-go-name": "weekendItem",
      "x-go-package": "polkovnik/api"
    },
    "WeekendInterval": {
      "type": "object",
      "properties": {
        "end": {
          "description": "Дата окончания выходного (включительно до конца дня).",
          "type": "string",
          "x-go-name": "End"
        },
        "start": {
          "description": "Дата начала выходного",
          "type": "string",
          "x-go-name": "Start"
        }
      },
      "x-go-name": "weekendInterval",
      "x-go-package": "polkovnik/api"
    }
  },
  "responses": {
    "ResponseError": {
      "description": "Стандартное сообщение об ошибке.",
      "headers": {
        "error": {
          "type": "string",
          "description": "Сообщение об ошибке"
        }
      }
    },
    "ResponseSuccess": {
      "description": "Стандартное сообщение об успешном выполнении запроса.",
      "headers": {
        "result": {
          "type": "string",
          "description": "Результат"
        }
      }
    },
...
    "User": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/User"
      }
    },
    "UserList": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/Users"
      }
    }
  }
}